featurestore/requirements.txt
fastapi
uvicorn
boto3
s3fs
pandas
pyarrow
python-dotenv
